OGM-665 The RESTEasy download of modules script breaks on Windows
The issue was that it would download a zip file from an URL ending with "filename.zip/download",
and the plugin would attempt to save the resource using that filename.
The slash makes for an illegal filename in Windows, so I had to find an alternative link
which wouldn't use such a pattern.