OGM-767 Add flexibility to batch fetching ORM engine
Disabled all but one strategy for the MVP
Introduce ability to customize how the inner loader is built
as OGM and ORM have different loaders.
Introduce notion of BatchableEntityLoader to abstract OgmLoader,
the legacy EntityLoader and plan.EntityLoader.
Methods of this interface is to be moved to UniqueEntityLoader
when moving things back to ORM.