OGM-947 Implement Redis Cluster support
Add support for Redis Cluster. Upgrade lettuce driver version. Document how to test OGM Redis with Redis Standalone and Redis Cluster. Create Makefile for Redis setup. Create a specific forEachTuple test since the cluster scan requires a scan over all master nodes.

The cluster routing is handled by the lettuce driver library, so mainly the interface changed to a smaller type which is implemented by the Cluster and the Standalone API. The lettuce driver cannot handle a cluster-wide SCAN command that is required to iterate over all keys in the Redis keyspace. This is done within the Redis dialect code and requires to keep track of the iteration state.