OGM-975 Extends the driver.* properties to support object retrievers
The hibernate.ogm.mongodb.driver.* which take an object were not supported so far.
Now it is possible to provide as value a String with a user-provided class name.
That class must implement stayic methods with the same name the methods have in
MongoClientOption.Builder.Builder. Those methods must return objects of the
correponding class.