OGM-1103 Initialize the states at the beginning of the event cycle
Prior to this patch, elements might have been created in the states
map while iterating on the map.

Typically, when all the operations are batched, the operations are
stored in the OperationCollector in the onFinish call for the
OperationsQueue. If all the operations are batched then the
OperationCollector was not initialized before that and we end up with a
ConcurrentModificationException.