OGM-1113 Ensure settings passed in the configure method of a test are correctly applied
So basically, it was already working but the test machinery does not
make it very clear. There were quite a lot of leftovers from various
iterations and the main objective of this commit is to make everything
more clear so that we understand what is going on.

I mostly renamed classes and interfaces so that they clearly state what
they are and to improve consistency. Typically, TestableGridDialect
must have been called that way when the plan was to integrate it in the
dialect but actually it is a GridDialectTestHelper.

Also renamed sessions to sessionFactory. And so on.