OGM-1064 Separate TupleContext in TupleContext and TupleTypeContext
This is similar to what was already done for AssociationContext and
AssociationTypeContext: it is cleaner and necessary for commits to
follow.

A following commit will show how it currently can hide bugs due to the
fact that we don't separate the context of an operation from the one
that can be reused.

Also create a common interface OperationContext for TupleContext and
AssociationContext.