OGM-1064 Clear the association after creation if not empty
In the case of an association stored in the document, while we are
creating a new empty association in createAssociation, we might end up
with a not empty association as the snapshot document might contain the
original entries. Thus, we add a clear in createAssociation to be sure
the created association is empty.