OGM-1064 Flush pending operations in case of generated properties
When properties are generated by the datastore, we need to flush the
pending operations to the datastore to be able to process the generated
properties.

It is especially important in the case of CouchDB as the revision of
each document is a property generated by the datastore.

Also introduce AbstractGroupingByEntityDialect to share code between
grouping aware dialects.