OGM-1229 Remove the EntityManager initialization branch as (Ogm)SessionImpl is now an EntityManager
Also removes a lot of copied code to be able to determine what is
strictly necessary. Note that this code will be reintroduced later (and
then in the end removed due thanks to some changes in ORM).

Note that this commit removes some APIs (NoSQLQuery and
NoSQLQueryImplementor) as they are not necessary anymore and requires us
to maintain more things.

We discussed it with Sanne and, considering the fact that ORM 5.2 breaks
a lot of things anyway, it would be considered acceptable.