OGM-1359 Clarify the output of the Travis build
1. If checkstyle fails, we don't want Travis to tell us the build
"errored": it just failed. Thus the checkstyle check should be in the
"script" section, not in the "install" section.
2. Running multiple independent script commands leads to an unclear
output, because the commands are executed regardless of the results of
the previous one. This means you can end up with build logs ending with
SUCCESS, while the build itself is marked as failed since one of the
commands failed.
   To work around the issue, I made sure commands are executed
conditionally, only if the previous one succeeded.
3. When checkstyle failed, it used to fail in the initial, no-log
build, and then we didn't have the checkstyle logs to tell us why it
failed. By disabling checkstyle in this initial build we can make the
build fail a little bit later, when the logs are enabled.