OGM-1363 Remove internal classes (and only internal classes) from the javadoc
We used to exclude protected members, which may be part of the API,
for instance in abstract classes.

This changes the filters to include all members, including protected
ones, but only those that are not in a 'impl' package. The result is a
view of all classes and members that Hibernate OGM users should normally
access.